app/requirements.txt
# Usage: pip install -r requirements.txt
# @TODO: add developer-specific requirements


# Flask Framework
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.2
MySQL-python==1.2.5
WTForms==2.0.2

# == Shibboleth http://flask-sso.readthedocs.org/en/latest/#a-minimal-example
#flask-sso

# for development
Flask-DebugToolbar==0.9.2
pytest==2.7.0
pytest-cov==1.8.1
Flask-Testing==0.4.2
nose==1.3.6

